fix settings
