+++ next.config.js
@@ -3,7 +3,7 @@ module.exports = {
   env: {
     // Reference a variable that was defined in the .env file and make it available at Build Time
     GA_TRACKING_ID: process.env.GA_TRACKING_ID,
-    BACKEND_URL: process.env.BLOG_BACKEND_URL,
+    BLOG_BACKEND_URL: process.env.BLOG_BACKEND_URL,
   },
   webpack: config => {
     config.node = {
