+++ src/pages/_app.tsx
@@ -1,11 +1,5 @@
 import 'ress';
 
-import { InMemoryCache } from 'apollo-cache-inmemory';
-import ApolloClient from 'apollo-client';
-import { ApolloLink } from 'apollo-link';
-import { onError } from 'apollo-link-error';
-import { HttpLink } from 'apollo-link-http';
-import fetch from 'isomorphic-unfetch';
 import App from 'next/app';
 import Head from 'next/head';
 import React from 'react';
@@ -14,6 +8,7 @@ import { createGlobalStyle } from 'styled-components';
 import { ApolloProvider } from '@apollo/react-hooks';
 
 import { NavBar } from '../containers/NavBar';
+import { client } from '../utils/apollo';
 
 const GlobalStyle = createGlobalStyle`
   html {
@@ -104,14 +99,6 @@ const GlobalStyle = createGlobalStyle`
 `;
 export default class extends App {
   render() {
-    const httpLink = new HttpLink({ uri: process.env.BLOG_BACKEND_URL, fetch: fetch });
-    const cache = new InMemoryCache();
-    const errorLink = onError(({ graphQLErrors }) => {
-      if (graphQLErrors) graphQLErrors.map(({ message }) => console.log(message));
-    });
-
-    const link = ApolloLink.from([errorLink, httpLink]);
-    const client = new ApolloClient({ link, cache });
     const { Component, pageProps } = this.props;
     return (
       <React.Fragment>
